% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODRI.R
\name{qrnegloglikensum}
\alias{qrnegloglikensum}
\title{Negative log-likelihood for Ordinal Models with more than 3 outcomes}
\usage{
qrnegloglikensum(deltain, y, x, beta, p)
}
\arguments{
\item{deltain}{initialization of cut-points.}

\item{y}{dependent variable i.e. ordinal outcome values.}

\item{x}{covariate matrix of dimension \eqn{(n x k)} including a column of ones.}

\item{beta}{column vector of coefficients of dimension \eqn{(k x 1)}.}

\item{p}{quantile level or skewness parameter, p in (0,1).}
}
\value{
Returns a list with components
\itemize{
\item{\code{nlogl}: }{a vector with likelihood values.}
\item{\code{negsumlogl}: }{a scalar with value of negative log-likelihood.}
}
}
\description{
Function for calculating negative log-likelihood for Ordinal models with
more than 3 outcomes.
}
\details{
Computes the negative of the log-likelihood function using the
asymmetric Laplace distribution over the iid random
variables.
}
\examples{
set.seed(101)
deltain <- c(-0.002570995, 1.044481071)
data("data25j4")
x <- data25j4$x
y <- data25j4$y
p <- 0.25
beta <- c(0.3990094, 0.8168991, 2.8034963)
output <- qrnegloglikensum(deltain, y, x, beta, p)

# nlogl
#   2.28675506
#   0.88417292
#   1.23613687
#   0.93127471
#   0.93866993
#   0.97694495
#   1.04934754
#   1.57723130
#   0.19231567
#   0.96491711... soon

# negsumlogl
#   285.7158

}
\references{
Rahman, M. A. (2016). “Bayesian
Quantile Regression for Ordinal Models.”
Bayesian Analysis, 11(1): 1-24.
}
\seealso{
likelihood maximization
}
