% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODRI.R
\name{quan_regg3}
\alias{quan_regg3}
\title{Bayesian Quantile Regression for Ordinal Models
with more than 3 outcomes}
\usage{
quan_regg3(y, x, mc = 15000, p, tune = 0.1)
}
\arguments{
\item{y}{dependent variable i.e. ordinal outcome values.}

\item{x}{covariate matrix of dimension \eqn{(n x k)} including a column of ones with or without column names.}

\item{mc}{number of MCMC iterations, post burn-in.}

\item{p}{quantile level or skewness parameter, p in (0,1).}

\item{tune}{tuning parameter.}
}
\value{
Returns a list with components:
\itemize{
\item{\code{post_mean_beta}: }{a vector with mean of sampled
 \eqn{\beta} for each covariate.}
 \item{\code{post_mean_beta}: }{a vector with mean of sampled
 \eqn{\beta} for each covariate.}
 \item{\code{post_mean_delta}: }{a vector with mean of sampled
 \eqn{\delta} for each cut point.}
 \item{\code{post_std_beta}: }{a vector with standard deviation
 of sampled \eqn{\beta} for each covariate.}
 \item{\code{post_std_delta}: }{a vector with standard deviation
 of sampled \eqn{\delta} for each cut point.}
 \item{\code{gamma}: }{a vector of cut points including Inf and
 -Inf.}
 \item{\code{catt}}
 \item{\code{acceptance_rate}: }{a scalar to judge the acceptance
 rate of samples.}
 \item{\code{DIC_result}: }{results of the DIC criteria.}
 \item{\code{beta_draws}: }{a matrix with all sampled values for \eqn{\beta}.}
 \item{\code{delta_draws}: }{a matrix with all sampled values for \eqn{\delta}.}
}
}
\description{
This function estimates the Bayesian Quantile Regression for ordinal models with
more than 3 outcomes and reports the posterior mean and posterior standard deviations
of \eqn{(\beta, \delta)}.
}
\details{
Function implements the Bayesian quantile regression for
ordinal models with more than 3 outcomes using a combination of Gibbs sampling
procedure and Metropolis-Hastings algorithm.

Function initializes prior and then iteratively
samples \eqn{\beta}, \eqn{\delta} and latent variable z.
Burn-in is taken as \eqn{0.25*mc} and \eqn{N = burn}-\eqn{in + mc}.
}
\examples{
 set.seed(101)
 data("data25j4")
 x <- data25j4$x
 y <- data25j4$y
 p <- 0.25
 output <- quan_regg3(y, x, mc = 50, p, 0.89)


 # Number of burn-in draws: 12.5
 # Number of retained draws: 50
 # Summary of estimated beta:

 #          Bayes Estimate Std Value Upper Credible Lower Credible
 # beta_0          0.7669    0.7669         1.3640         0.0862
 # beta_1          0.9110    0.9110         1.2914         0.4157
 # beta_2          3.2143    3.2143         4.0231         2.0865
 # delta_1         0.6665    0.2718         0.9781         0.2646
 # delta_2         0.8351    0.3719         1.1934         0.0040

 # Model acceptance rate: 36
 # Model DIC: 732.6

}
\references{
Rahman, M. A. (2016). “Bayesian
Quantile Regression for Ordinal Models.”
Bayesian Analysis, 11(1): 1-24.

Yu, K. and Moyeed, R. A. (2001). “Bayesian Quantile Regression.” Statistics and
Probability Letters, 54(4): 437–447.

Casella, G., George E. I. (1992). “Explaining the Gibbs Sampler."
The American Statistician, 46(3): 167-174.

Geman, S., and Geman, D. (1984). “Stochastic Relaxation,
Gibbs Distributions, and the Bayesian Restoration of Images."
IEEE Transactions an Pattern Analysis and Machine Intelligence,
6(6): 721-741.

 Chib, S., Greenberg E. (1995). "Understanding the Metropolis-Hastings
 Algorithm." The American Statistician, 49(4): 327-335.

 Hastings, W.K. (1970). "Monte Carlo Sampling Methods Using
 Markov Chains and Their Applications." Biometrika, 57: 1317-1340.
}
\seealso{
tcltk, \link[stats]{rnorm}, \link[stats]{qnorm},
\link[MASS]{ginv}, Gibbs sampler
}
