% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORII.R
\name{devianceOR2}
\alias{devianceOR2}
\title{Deviance Information Criteria for ordinal quantile model
with 3 outcomes}
\usage{
devianceOR2(y, x, gammacp, p, postMeanbeta, postStdbeta,
postMeansigma, postStdsigma, beta, sigma, burn, nsim)
}
\arguments{
\item{y}{observed ordinal outcomes, column vector of dimension \eqn{(n x 1)}.}

\item{x}{covariate matrix of dimension \eqn{(n x k)} including a column of ones with or without column names.}

\item{gammacp}{row vector of cut-points including -Inf and Inf.}

\item{p}{quantile level or skewness parameter, p in (0,1).}

\item{postMeanbeta}{mean value of \eqn{\beta} obtained from MCMC draws.}

\item{postStdbeta}{standard deviation of \eqn{\beta} obtained from MCMC draws.}

\item{postMeansigma}{mean value of \eqn{\sigma} obtained from MCMC draws.}

\item{postStdsigma}{standard deviation of \eqn{\sigma} obtained from MCMC draws.}

\item{beta}{MCMC draw of coefficients, dimension is \eqn{(k x nsim)}.}

\item{sigma}{MCMC draw of scale factor, dimension is \eqn{(nsim x 1)}.}

\item{burn}{number of discarded MCMC iterations.}

\item{nsim}{total number of MCMC iterations including the burn-in.}
}
\value{
Returns a list with components
\deqn{DIC = 2*avgdeviance - devpostmean}
\deqn{pd = avgdeviance - devpostmean}
\deqn{devpostmean = -2*(logLikelihood)}.
}
\description{
Function for computing the Deviance information criteria for ordinal quantile
model with 3 outcomes.
}
\details{
Deviance is -2*(log likelihood) and has an important role in
statistical model comparison because of its relation with Kullback-Leibler
information criteria.
}
\examples{
set.seed(101)
data("data25j3")
x <- data25j3$x
y <- data25j3$y
k <- dim(x)[2]
output <- quantregOR2(y = y, x = x, b0 = 0, B0 = 10*diag(k), n0 = 5, d0 = 8, gamma = 3,
mcmc = 50, p = 0.25, verbose = FALSE)
gammacp <- c(-Inf, 0, 3, Inf)
postMeanbeta <- output$postMeanbeta
postStdbeta <- output$postStdbeta
postMeansigma <- output$postMeansigma
postStdsigma <- output$postStdsigma
beta <- output$beta
sigma <- output$sigma
mcmc = 50
burn <- 10
nsim <- burn + mcmc
deviance <- devianceOR2(y, x, gammacp, p = 0.25, postMeanbeta, postStdbeta,
postMeansigma, postStdsigma, beta, sigma, burn, nsim)

# DIC
#   801.8191
# pd
#   6.608594
# devpostmean
#   788.6019

}
\references{
Rahman, M. A. (2016). “Bayesian
Quantile Regression for Ordinal Models.”
Bayesian Analysis, 11(1): 1-24. DOI: 10.1214/15-BA939

Spiegelhalter, D. J., Best, N. G., Carlin, B. P. and Linde, A. (2002).
“Bayesian Measures of Model Complexity and Fit.” Journal of the
Royal Statistical Society B, Part 4: 583-639. DOI: 10.1111/1467-9868.00353

Gelman, A., Carlin, J. B., Stern, H. S., and Rubin, D. B.
“Bayesian Data Analysis.” 2nd Edition, Chapman and Hall. DOI: 10.1002/sim.1856
}
\seealso{
decision criteria
}
